feature: set unique id for satellites in TLE file. Add 'all' option in config file to track different kinds of satellites in TLE file
With 'all' option in track.ini, user can track all kinds of satellites in TLE file. For instace, if the TLE file has: [DElTA, ONEWEB, STARLINK, ...], if option all is used, then it will consider all satellites, otherwise it will track the provided satellite, in this case, the used could decide just for DELTA satellites. The unique identifier was implemented to distinguish entries in TLE file with the same name, but that refer to different objects, this way, 'all' option can track them all.